backend/requirements.txt
asgiref==3.2.10
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-extensions==3.0.8
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.11.1
graphviz==0.14.1
idna==2.10
jmespath==0.10.0
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
pilkit==2.0
Pillow==7.2.0
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
django-cleanup==5.1.0
